Renamed `useTemporaryQuery` to `useScopedQuery`
