+++ packages/@livestore/react/src/mod.ts
@@ -1,7 +1,7 @@
 export { LiveStoreContext, useStore } from './LiveStoreContext.js'
 export { LiveStoreProvider } from './LiveStoreProvider.js'
 export { useQuery } from './useQuery.js'
-export { useTemporaryQuery } from './useTemporaryQuery.js'
+export { useScopedQuery } from './useScopedQuery.js'
 export { useStackInfo } from './utils/stack-info.js'
 export {
   useRow,
