+++ packages/@livestore/react/src/useScopedQuery.ts
